无配置即可体验
