+++ key-mgr/src/main/java/com/webank/keymgr/config/db/MultiTransactionConfig.java
@@ -13,12 +13,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.webank.keymgr.config;
+package com.webank.keymgr.config.db;
 
 import com.atomikos.icatch.jta.UserTransactionImp;
 import com.atomikos.icatch.jta.UserTransactionManager;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnExpression;
 import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Conditional;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.DependsOn;
 import org.springframework.transaction.PlatformTransactionManager;
@@ -36,7 +37,7 @@ import javax.transaction.UserTransaction;
  * @data 2019-07-28 14:17:27
  *
  */
-@ConditionalOnExpression("'${system.mgrStyle:file}'.equals('db')")
+@ConditionalOnExpression("'db'.equals('${system.mgrStyle}')")
 @Configuration
 @EnableTransactionManagement
 public class MultiTransactionConfig {
