+++ key-mgr/src/main/java/com/webank/keymgr/db/dao/EncryptKeyInfoDAO.java
@@ -3,7 +3,7 @@
  */
 package com.webank.keymgr.db.dao;
 
-import com.webank.keymgr.config.EncryptKeyJpaConfig;
+import com.webank.keymgr.config.db.EncryptKeyJpaConfig;
 import com.webank.keymgr.config.SystemConfig;
 import com.webank.keymgr.db.encryptkey.entity.EncryptKeyDO;
 import com.webank.keymgr.db.encryptkey.repository.EncryptKeyInfoRepository;
